Move lib/ into ./src
