+++ src/api/mapping.ts
@@ -1,6 +1,6 @@
 import express from 'express'
 import uuid4 from 'uuid/v4'
-import { setData, getMappings } from './lib/data'
+import { setData, getMappings } from '../lib/data'
 import { Mapping } from './types/general'
 const mappingRouter = express.Router()
